+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 -->
 ## Unreleased
+### Added
+- [#94](https://github.com/CCSDForge/episciences-front/issues/94): New labels for boards
 ### Fixed
 - The number of publications with one author is incorrect because the search is done by Surname or First name instead of First name and Surname.
 - Wrong statistical indicator "Reviews received".